Fix Discount function

+++ OoadProject.Core/ViewModels/Sells/Dtos/InvoiceForCreationDto.cs
@@ -11,16 +11,16 @@ namespace OoadProject.Core.ViewModels.Sells.Dtos
         private string _customerName;
         private string _phoneNumber;
         private int _total;
-        private float _discount;
-        private float _price;
+        private int _discount;
+        private int _price;
         private string _customerLevel;
 
 
         public string CustomerName { get => _customerName; set { _customerName = value; OnPropertyChanged(); } }
         public string PhoneNumber { get => _phoneNumber; set { _phoneNumber = value; OnPropertyChanged(); } }
         public int Total { get => _total; set { _total = value; OnPropertyChanged(); } }
-        public float Discount { get => _discount; set { _discount = value; OnPropertyChanged(); } }
-        public float Price { get => _price; set { _price = value; OnPropertyChanged(); } }
+        public int Discount { get => _discount; set { _discount = value; OnPropertyChanged(); } }
+        public int Price { get => _price; set { _price = value; OnPropertyChanged(); } }
         public string CustomerLevel { get => _customerLevel; set { _customerLevel = value; OnPropertyChanged(); } }
 
     }
